billing: Revert value to int, not float (#244)
This was part of #172 that I'd mistakenly suggested. It turns out the
control plane actually cannot accept fractional values for billing
values.

Validated with a shim billing implementation, numbers look right.
Merged without review so that staging isn't left broken.

+++ pkg/billing/model.go
@@ -21,5 +21,5 @@ type IncrementalEvent struct {
 	EndpointID     string    `json:"endpoint_id"`
 	StartTime      time.Time `json:"start_time"`
 	StopTime       time.Time `json:"stop_time"`
-	Value          float64   `json:"value"`
+	Value          int       `json:"value"`
 }
